StoryTest1: Hook up tests in cap_test to test cap-protocol.c implementation  * Fix minor bugs found as a result:     - cap_make_connection(): close the FD if no connection is created.     - plpy_make_conn2(): decrement reference counts properly.     - plpy_obj_hash(): use correct return type: long instead of int, which       makes a difference on 64-bit systems  * comms_test.py:  Minor refactor.  Add copyright message.
git-svn-id: svn://svn.gna.org/svn/plash/trunk@680 e2ebe5e9-4d0c-0410-8560-844d39661dae

+++ python/py-type-obj.c
@@ -141,9 +141,9 @@ static PyObject *plpy_compare(plpy_obj *obj1, PyObject *obj2)
 }
 */
 
-static int plpy_obj_hash(plpy_obj *self)
+static long plpy_obj_hash(plpy_obj *self)
 {
-  return (int) self->obj;
+  return (long) self->obj;
 }
